+++ src/protect.cpp
@@ -47,6 +47,11 @@ SEXP wrapped_mkchar_utf8(void *data) {
   return Rf_mkCharCE(*c, CE_UTF8);
 }
 
+SEXP wrapped_mkchar_len_utf8(void *data) {
+  struct safe_mkchar_len_data *c = (struct safe_mkchar_len_data*) data;
+  return Rf_mkCharLenCE(c->c, c->len, CE_UTF8);
+}
+
 SEXP wrapped_mkstring(void *data) {
   const char **c = (const char **) data;
   return Rf_mkString(*c);